Organazing refresh token and user code on domain
