--- internal/domain/datasource/cache/user.go
@@ -1,15 +1,0 @@
-package domain
-
-import (
-	"context"
-
-	"github.com/daniarmas/notes/internal/domain/entity"
-	"github.com/google/uuid"
-)
-
-type UserCacheDs interface {
-	GetUser(ctx context.Context, id uuid.UUID) (*entity.User, error)
-	CreateUser(ctx context.Context, user *entity.User) (*entity.User, error)
-	UpdateUser(ctx context.Context, user *entity.User) (*entity.User, error)
-	DeleteUser(ctx context.Context, id uuid.UUID) error
-}
