+++ internal/domain/refresh_token.go
@@ -1,4 +1,4 @@
-package entity
+package domain
 
 import (
 	"time"
