+++ internal/domain/refresh_token_cache_ds.go
@@ -1,0 +1,13 @@
+package domain
+
+import (
+	"context"
+
+	"github.com/google/uuid"
+)
+
+type RefreshTokenCacheDs interface {
+	GetRefreshToken(ctx context.Context, id uuid.UUID) (*RefreshToken, error)
+	CreateRefreshToken(ctx context.Context, refreshToken *RefreshToken) (*RefreshToken, error)
+	DeleteRefreshToken(ctx context.Context, id uuid.UUID) error
+}
