+++ internal/domain/refresh_token_repository.go
@@ -1,0 +1,13 @@
+package domain
+
+import (
+	"context"
+
+	"github.com/google/uuid"
+)
+
+type RefreshTokenRepository interface {
+	GetRefreshToken(ctx context.Context, id uuid.UUID) (*RefreshToken, error)
+	CreateRefreshToken(ctx context.Context, refreshToken *RefreshToken) (*RefreshToken, error)
+	DeleteRefreshToken(ctx context.Context, id uuid.UUID) error
+}
