+++ internal/domain/user_cache_ds.go
@@ -1,0 +1,14 @@
+package domain
+
+import (
+	"context"
+
+	"github.com/google/uuid"
+)
+
+type UserCacheDs interface {
+	GetUser(ctx context.Context, id uuid.UUID) (*User, error)
+	CreateUser(ctx context.Context, user *User) (*User, error)
+	UpdateUser(ctx context.Context, user *User) (*User, error)
+	DeleteUser(ctx context.Context, id uuid.UUID) error
+}
